Added validation of job input files to compose
