+++ wrfhydropy/__init__.py
@@ -3,4 +3,5 @@ from .core.job import *
 from .core.model import *
 from .core.simulation import *
 from .core import schedulers
-from .core import fileutilities
+from .core import outputdiffs
+from .core import ioutils
