+++ wrfhydropy/core/domain.py
@@ -3,7 +3,7 @@ import json
 import re
 import shutil
 
-from .fileutilities import \
+from .ioutils import \
     WrfHydroStatic, \
     WrfHydroTs
