+++ wrfhydropy/core/job.py
@@ -10,7 +10,7 @@ import pickle
 import os
 import copy
 
-from .fileutilities import check_file_exist_colon
+from .ioutils import _check_file_exist_colon
 
 class Job(object):
     def __init__(
@@ -119,12 +119,12 @@ class Job(object):
         # Check for restart files both as specified and in the run dir..
         # Alias the mutables for some brevity
         hydro_nlst = self.hydro_namelist['hydro_nlist']
-        hydro_nlst['restart_file'] = check_file_exist_colon(current_dir, hydro_nlst[
+        hydro_nlst['restart_file'] = _check_file_exist_colon(current_dir, hydro_nlst[
             'restart_file'])
         nudging_nlst = self.hydro_namelist['nudging_nlist']
         if nudging_nlst:
             nudging_nlst['nudginglastobsfile'] = \
-                check_file_exist_colon(current_dir, nudging_nlst['nudginglastobsfile'])
+                _check_file_exist_colon(current_dir, nudging_nlst['nudginglastobsfile'])
 
         # Copy namelists from job_dir to current_dir
         hydro_namelist_path = self.job_dir.joinpath('hydro.namelist')
@@ -224,8 +224,8 @@ class Job(object):
         # This is needed because the model outputs restarts with colons, and our distributed
         # domains do not have restarts with colons so that they can be easily shared across file
         # systems
-        hydro_restart_file = check_file_exist_colon(os.getcwd(),hydro_restart_basename)
-        nudging_restart_file = check_file_exist_colon(os.getcwd(),nudging_restart_basename)
+        hydro_restart_file = _check_file_exist_colon(os.getcwd(),hydro_restart_basename)
+        nudging_restart_file = _check_file_exist_colon(os.getcwd(),nudging_restart_basename)
 
         self.hydro_times['hydro_nlist']['restart_file'] = hydro_restart_file
         self.hydro_times['nudging_nlist']['nudginglastobsfile'] = nudging_restart_file
